Init
This is the initial work of the personal website, squashed old commits and the stuff I was learning from following along with a blog.

Added a gemfile and tried to get a theme working, but no dice on that; paginate needs to be fixed as well.

Also trying to figure out CSS and stylizing everything. That's a 'over the summer' project.