cv two columns
after further testing and implementation, inline css will handle positioning of columns if the viewer size is that of a mobile device as defined by the bulma framework.

also pushed into new html files the information for each column so that I can more visibly and easily edit the different sections in the future.