about-me and other things
worked on my about me section for the github profile read-me; I went ahead and added a author ag to the _config.yml file for inclusion as the site.author. A number of other changes will have to shortly follow to verify that it works.

proof-reading of blog posts ---

there were a few issues with the blog posts. hopefully some of the changes made implements properly through the jekyll on github's end